// src/app/services/models/data-response-page-order-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PageOrderDto } from '../models/page-order-dto';

/**
 * Generic data response
 */
export interface DataResponsePageOrderDto {
  results?: PageOrderDto;

  /**
   * Status code of the operation
   */
  statusCode?: string;

  /**
   * Detailed status message
   */
  statusMessage?: string;

  /**
   * Indicates if the operation succeeded
   */
  succeeded?: boolean;
}